Begin writing DBPF module
Unfortunately, a lot of the Python (and other languages) modules I found
in the wild did not work (many years or a decade old) or were not fit
for purpose.

So, I decided to give it a whirl and implement the DBPF spec so this
script can extract and pack the ui.package file later.