Generate a .package file with new files
Based on the structure produced by SimPE, which are XMLs describe the
Group ID and Instance ID.

The header was checked byte for byte compared to a new package produced
by SimPE, to check it works. Compared to the original, the file is large
because compression is not implemented.

At time of commit, this does not work with the game! (It crashes after
the splash video, and destroys ui.package if it has write access)