Port patcher program to PyQt6
I liked TKinter for its lightweightness, but it ends up looking
inconsistent between Windows, Linux. Kind of ironic for a HiDPI patcher.

Qt is a framework with higher level features that makes it more suitable
for things like threads and will look native on the platform it's ran
on. The size of the distributed build will increase slightly, but that
should be OK!