Use multiple processes to speed up patching
Parallelisation will speed up the patching time. Refactor the logic to
support this and keep the UI updated on the main thread.

This breaks the CTRL+C signal, due to the Qt event loop and
multiprocesses, so the application must be closed via GUI if the patch
is in progress, otherwise the processes just 'stop' which is worse.