DBPF: Optimise statefulness for existing packages
Refactor and optimise the module so we load, decompress and compress
data when needed. Enables new packages to be built in a way that
retains the existing compression if the original can't be decompressed.

This prepares the module for processing a package like objects.package
which has many unknown data types which we couldn't decompress. It
also happens to fix the DIR entry's group and instance IDs that were
previously 0, which may (or may not) have mattered.

More efficient then the earlier implementation (fcde15ca2f64)
which extracted the entire package contents into memory first!