Refactor GameFile class and patching mechanism
Enable our logic to support two kinds of patching:

- GameFileReplacement: The file is swapped for our patched version.
- GameFileOverride: Save differences to a new package in "Overrides".

For the latter (GameFileOverride), backup files are not necessary, since
the original package is always left intact.

This prepares support for "objects.package", which is a huge package
compressed weirdly, and quite an important file to be messing with.
Instead, save the differences in the "Overrides" folder, as opposed
to attempting to repack that package and crashing the game.

The specific files to be patched were moved to gamefile.py and
patches.py, so the patcher program is more abstract.

Drops the unused MD5 integrity checksum added in dd7e863afd8b.