Tools/Inspector: Inspect all UI packages from game
Instead of reading a single package, use the game directory to locate
all UI scripts, including CaSIEUI.data.

The tree lists unique instances of each group/instance, and indicates
the game(s) and package(s) where the UI script can be found. This does
cause the tool to load a bit slower, but gives us more to work with!