UIScript: Add deserialization methods
The game has some duplicate attributes (wparam, style). It seems like
the last instance of the attribute might be used, so we'll assume that
for now.

Duplicate attributes happens in IDs like:

- 0x49001057: "Moving Family" or "Moving Household" (bitmap/text)
- 0x8c159251: Buy Clothing (there's two icons sets due to "Young Adult")
- 0x8c159280: "Change Appearance"
- 0x8c159286: "Perform Plastic Survey" (never knew that was a thing)
- 0xfffffff0: Skin browser (debug)
- 0xfffffff1: Outfit browser (debug)

There are also other inconsistencies, like multiline captions and an
attribute like "initvalue" that can sometimes, but always, be quoted.
In some files, it has an equals but nothing after it ("initvalue=")