UIScript: Support duplicate attributes
Duplicate attributes in elements isn't XML compliant, but it seems
the game does need multiple "wparam" attributes as it uses them to
swap buttons/bitmaps depending on which EPs are installed.

Support this by introducing an optional list data type for attributes.

The game expects the attributes to be placed next to each other.